+++ src/pykiso/lib/auxiliaries/record_auxiliary.py
@@ -20,7 +20,6 @@ Record Auxiliary
 
 import io
 import logging
-import multiprocessing
 import re
 import sys
 import threading
@@ -29,26 +28,21 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 from pykiso import CChannel
-from pykiso.interfaces.dt_auxiliary import (
-    DTAuxiliaryInterface,
+from pykiso.auxiliary import (
+    AuxiliaryInterface,
     close_connector,
     open_connector,
 )
+from pykiso.lib.connectors.cc_proxy import CCProxy
 
 log = logging.getLogger(__name__)
 
 
 class StringIOHandler(io.StringIO):
     def __init__(self, multiprocess: bool = False) -> None:
-        """Constructor
-
-        :param multiprocess: use a thread or multiprocessing lock.
-        """
+        """Constructor"""
         super(StringIOHandler, self).__init__()
-        if multiprocess:
-            self.data_lock = multiprocessing.Lock()
-        else:
-            self.data_lock = threading.Lock()
+        self.data_lock = threading.Lock()
 
     def get_data(self) -> str:
         """Get data from the string
@@ -67,7 +61,7 @@ class StringIOHandler(io.StringIO):
             self.write(data)
 
 
-class RecordAuxiliary(DTAuxiliaryInterface):
+class RecordAuxiliary(AuxiliaryInterface):
     """Auxiliary used to record a connectors receive channel."""
 
     LOG_HEADER = "Received data :"
@@ -91,11 +85,7 @@ class RecordAuxiliary(DTAuxiliaryInterface):
         :param timeout: timeout for the receive channel
         :param log_path: path to the log folder
         :param max_file_size: maximal size of the data string
-        :param multiprocess: use a Process instead of a Thread for
-            active polling.
-            Note1: the data will automatically be saved.
-            Note2: if proxy usage, all connectors should be 'CCMpProxy'
-            and 'processing' flag set to True
+        :param multiprocess: deprecated, will not be taken into account.
         :param manual_start_record: flag to not start recording on
             auxiliary creation
         """
@@ -107,19 +97,14 @@ class RecordAuxiliary(DTAuxiliaryInterface):
         self.timeout = timeout
         self.stop_receive_event = None
         self._receive_thread_or_process = None
-        self.multiprocess = multiprocess
         self.cursor = 0
         self.log_folder_path = log_folder_path
+        self.multiprocess = multiprocess
         self._data = StringIOHandler(multiprocess)
         self.max_file_size = max_file_size
 
         if self.is_active and not manual_start_record:
             self.start_recording()
-
-        if self.multiprocess:
-            log.internal_warning(
-                "Logs will only be dumped into a file due due to the multiprocess flag"
-            )
 
     def get_data(self) -> str:
         """Return the entire log buffer content.
@@ -200,10 +185,6 @@ class RecordAuxiliary(DTAuxiliaryInterface):
                 else:
                     self.set_data("\n" + stream)
 
-        if self.multiprocess:
-            # dump data inside the process
-            self.dump_to_file(f"record_{type(self.channel).__name__}.log")
-
         try:
             self.channel.close()
         except Exception:
@@ -235,7 +216,7 @@ class RecordAuxiliary(DTAuxiliaryInterface):
     def clear_buffer(self) -> None:
         """Clean the buffer that contain received messages."""
         log.internal_info("Clearing buffer")
-        self._data = StringIOHandler(self.multiprocess)
+        self._data = StringIOHandler()
         self.cursor = 0
 
     def stop_recording(self) -> None:
@@ -257,15 +238,9 @@ class RecordAuxiliary(DTAuxiliaryInterface):
             self._receive_thread_or_process is None
             or not self._receive_thread_or_process.is_alive()
         ):
-            # define multiprocessing or Thread variables
-            if self.multiprocess:
-                self.stop_receive_event = multiprocessing.Event()
-                self._receive_thread_or_process = multiprocessing.Process(
-                    target=self.receive
-                )
-            else:
-                self.stop_receive_event = threading.Event()
-                self._receive_thread_or_process = threading.Thread(target=self.receive)
+            # define Thread variables
+            self.stop_receive_event = threading.Event()
+            self._receive_thread_or_process = threading.Thread(target=self.receive)
 
             self.clear_buffer()
             self._receive_thread_or_process.start()