+++ tests/test_robot_instrument_control_auxiliary.py
@@ -9,7 +9,7 @@
 
 import pytest
 
-from pykiso.interfaces.thread_auxiliary import AuxiliaryInterface
+from pykiso.auxiliary import AuxiliaryInterface
 from pykiso.lib.auxiliaries.instrument_control_auxiliary.lib_scpi_commands import (
     LibSCPI,
 )
@@ -44,9 +44,7 @@ class MockVisaChannel:
 def instrument_aux_instance(
     mocker,
 ):
-    mocker.patch(
-        "pykiso.interfaces.thread_auxiliary.AuxiliaryInterface.start", return_value=None
-    )
+    mocker.patch("pykiso.auxiliary.AuxiliaryInterface.start", return_value=None)
     mocker.patch(
         "pykiso.test_setup.config_registry.ConfigRegistry.get_auxes_by_type",
         return_value={"inst_aux": InstAux(MockVisaChannel(), "instrument", 1)},